\name{cmshcc-package}
\alias{cmshcc-package}
\alias{cmshcc}
\docType{package}
\title{
\packageTitle{cmshcc}
}
\description{
\packageDescription{cmshcc}
}
\details{

The DESCRIPTION file:
\packageDESCRIPTION{cmshcc}
\packageIndices{cmshcc}
~~ An overview of how to use the package, including the most important functions ~~
}
\author{
\packageAuthor{cmshcc}

Maintainer: \packageMaintainer{cmshcc}
}
\references{
~~ Literature or other references for background information ~~
}
~~ Optionally other standard keywords, one per line, from file KEYWORDS in the R ~~
~~ documentation directory ~~
\keyword{ package }
\seealso{
~~ Optional links to other man pages, e.g. ~~
~~ \code{\link[<pkg>:<pkg>-package]{<pkg>}} ~~
}
\examples{
~~ simple examples of the most important functions ~~
}
