% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{generate_DIAG}
\alias{generate_DIAG}
\title{Generate sample diagnosis data}
\usage{
generate_DIAG(cmshcc_map, size = 100, seed = 2, max_dx = 10)
}
\arguments{
\item{cmshcc_map}{data frame is the CMSHCC diagnosis to HCC mapping in question (Required)}

\item{size}{integer is the number of rows of diagnoses that are required (default 100)}

\item{seed}{integer is the random seed starting value for reproducability (default 2)}

\item{max_dx}{integer is the maximum number of diagnoses that a beneficiary can have (default 10)}
}
\value{
data frame DIAG is the list of HICNO and ICD diagnoses
}
\description{
Generate sample diagnosis data
}
