% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{generate_PERSON}
\alias{generate_PERSON}
\title{generateTestPERSON}
\usage{
generate_PERSON(size = 100, seed = 2, start_time = "1930/01/01",
  end_time = Sys.Date())
}
\arguments{
\item{size}{integer is the number of dates to simulate (default is 100)}

\item{seed}{integer is the random seed starting point}

\item{start_time}{string is the YYYY/MM/DD start time to start the sampled random dates (default is 1930/01/01)}

\item{end_time}{string is the YYYY/MM/DD end time to cut off the sampled random dates (default is today's date)}
}
\value{
Data Frame for the PERSON demographics information required for risk adjustment
}
\description{
generateTestPERSON
}
