% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{load_cmshcc_map}
\alias{load_cmshcc_map}
\title{load_hcc_map}
\usage{

  load_cmshcc_map(file_name = "data/2017MidyearFinalICD-10-CMMappings/2017_Midyear_Final ICD-10-CM Mappings_standard.csv")
}
\arguments{
\item{string}{file name for the ICD -> HCC mapping file}
}
\value{
Data Frame The data frame of the mapping file
}
\description{
load_hcc_map
}
