% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{person_age}
\alias{person_age}
\title{person_age}
\usage{
person_age(DOB, ASOF = Sys.Date())
}
\arguments{
\item{DOB}{date is the date of birth for the person}

\item{ASOF}{date is the date as of the risk scores are evaluated}
}
\value{
numeric age
}
\description{
person_age
}
