% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evaluate.r
\name{person_age_band}
\alias{person_age_band}
\title{person_age_band}
\usage{
person_age_band(ages, breaks = breaks_def, break_labels = break_labels_def)
}
\arguments{
\item{breaks}{vector of integers representing the lower endpoints for the age bands}

\item{break_labels}{vector of strings representing the labels for each age band}
}
\description{
person_age_band
}
